requirements.txt
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.7
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.8.0
google-auth==2.3.0
idna==3.2
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.1.0
pytz==2021.1
requests==2.26.0
rsa==4.7.2
sqlparse==0.4.2
urllib3==1.26.7
